+++ test/hardhat/helper/common.ts
@@ -4,14 +4,7 @@ import chaiAsPromised from 'chai-as-promised';
 import { config as dotenvConfig } from 'dotenv';
 import { ethers } from 'hardhat';
 
-import {
-  getCurrencyAmount,
-  getToken,
-  getUnwrapETHTx,
-  getWrapETHTx,
-} from '../../helper';
-import { ApertureSupportedChainId } from '../../interfaces';
-import { WETH__factory } from '../../typechain-types';
+import { ApertureSupportedChainId } from '../../../interfaces';
 
 dotenvConfig();
 
@@ -41,35 +34,3 @@ export async function resetHardhatNetwork() {
     /*blockNumber=*/ 17188000,
   );
 }
-
-describe('WETH transaction tests', function () {
-  beforeEach(async function () {
-    await resetHardhatNetwork();
-  });
-  it('Deposit and withdraw WETH', async function () {
-    const wethContract = WETH__factory.connect(
-      WETH_ADDRESS,
-      hardhatForkProvider,
-    );
-    const wethBalanceBefore = await wethContract.balanceOf(WHALE_ADDRESS);
-    const WETH = await getToken(WETH_ADDRESS, chainId, hardhatForkProvider);
-    const wrapAmount = getCurrencyAmount(WETH, '10').quotient.toString();
-    const whaleSigner = await ethers.getImpersonatedSigner(WHALE_ADDRESS);
-    await (
-      await whaleSigner.sendTransaction(getWrapETHTx(chainId, wrapAmount))
-    ).wait();
-    expect(
-      (await wethContract.balanceOf(WHALE_ADDRESS)).eq(
-        wethBalanceBefore.add(wrapAmount),
-      ),
-    ).to.equal(true);
-    await (
-      await whaleSigner.sendTransaction(getUnwrapETHTx(chainId, wrapAmount))
-    ).wait();
-    expect(
-      (await wethContract.balanceOf(WHALE_ADDRESS)).eq(
-        wethBalanceBefore.toString(),
-      ),
-    ).to.equal(true);
-  });
-});
